+++ SparkZBaseBuilding/scripts/4_World/Actions/SPKZ_DismantleWoodWall.c
@@ -40,6 +40,5 @@ modded class Screwdriver
  {
   super.SetActions();
   AddAction(SPKZ_ActionDismantleWoodWall);
-  AddAction(SPKZ_ActionDismantleWorkbench);
  }
 }