+++ SparkZBaseBuilding/scripts/4_World/Actions/SPKZ_WoodDoorActions.c
@@ -69,6 +69,5 @@ modded class ActionConstructor
   actions.Insert(SPKZ_ActionOpenWoodGarage);
   actions.Insert(SPKZ_ActionCloseWoodGarage);
   actions.Insert(SPKZ_ActionAccessWorkbench);
-  actions.Insert(SPKZ_ActionDismantleWorkbench);
  }
 }