+++ SparkZBaseBuilding/workbench_slots.hpp
@@ -1,0 +1,23 @@
+class CfgSlots
+{
+ class Slot_SPKZ_WB_Hacksaw {name="SPKZ_WB_Hacksaw";displayName="Hacksaw";ghostIcon="set:spkz_workbench image:Hacksaw";selection="SPKZ_WB_Hacksaw";show="true";};
+ class Slot_SPKZ_WB_HandSaw {name="SPKZ_WB_HandSaw";displayName="Handsaw";ghostIcon="set:spkz_workbench image:HandSaw";selection="SPKZ_WB_HandSaw";show="true";};
+ class Slot_SPKZ_WB_Hammer {name="SPKZ_WB_Hammer";displayName="Hammer";ghostIcon="set:spkz_workbench image:Hammer";selection="SPKZ_WB_Hammer";show="true";};
+ class Slot_SPKZ_WB_Shovel {name="SPKZ_WB_Shovel";displayName="Shovel";ghostIcon="set:spkz_workbench image:Shovel";selection="SPKZ_WB_Shovel";show="true";};
+ class Slot_SPKZ_WB_Screwdriver {name="SPKZ_WB_Screwdriver";displayName="Screwdriver";ghostIcon="set:spkz_workbench image:Screwdriver";selection="SPKZ_WB_Screwdriver";show="true";};
+ class Slot_SPKZ_WB_Pliers {name="SPKZ_WB_Pliers";displayName="Pliers";ghostIcon="set:spkz_workbench image:Pliers";selection="SPKZ_WB_Pliers";show="true";};
+ class Slot_SPKZ_WB_SledgeHammer {name="SPKZ_WB_SledgeHammer";displayName="Sledgehammer";ghostIcon="set:spkz_workbench image:SledgeHammer";selection="SPKZ_WB_SledgeHammer";show="true";};
+ class Slot_SPKZ_WB_SharpeningStone {name="SPKZ_WB_SharpeningStone";displayName="Sharpening stone";ghostIcon="set:spkz_workbench image:SharpeningStone";selection="SPKZ_WB_SharpeningStone";show="true";};
+};
+class CfgNonAIVehicles
+{
+ class ProxyAttachment;
+ class ProxySPKZ_WB_Hacksaw: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_Hacksaw.p3d";inventorySlot="SPKZ_WB_Hacksaw";};
+ class ProxySPKZ_WB_HandSaw: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_HandSaw.p3d";inventorySlot="SPKZ_WB_HandSaw";};
+ class ProxySPKZ_WB_Hammer: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_Hammer.p3d";inventorySlot="SPKZ_WB_Hammer";};
+ class ProxySPKZ_WB_Shovel: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_Shovel.p3d";inventorySlot="SPKZ_WB_Shovel";};
+ class ProxySPKZ_WB_Screwdriver: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_Screwdriver.p3d";inventorySlot="SPKZ_WB_Screwdriver";};
+ class ProxySPKZ_WB_Pliers: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_Pliers.p3d";inventorySlot="SPKZ_WB_Pliers";};
+ class ProxySPKZ_WB_SledgeHammer: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_SledgeHammer.p3d";inventorySlot="SPKZ_WB_SledgeHammer";};
+ class ProxySPKZ_WB_SharpeningStone: ProxyAttachment {scope=2;model="SparkZBaseBuilding\Data\Workbench\proxies\SPKZ_WB_SharpeningStone.p3d";inventorySlot="SPKZ_WB_SharpeningStone";};
+};